_archived/backend_archived/app/core/auth.py
"""Authentication utilities for Supabase JWT verification."""
from dataclasses import dataclass
from fastapi import Depends, HTTPException, status
from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
from jose import jwt

from .config import get_settings

settings = get_settings()
bearer_scheme = HTTPBearer(auto_error=False)


@dataclass
class User:
    """Authenticated user."""
    id: str


async def get_current_user(
    credentials: HTTPAuthorizationCredentials = Depends(bearer_scheme),
) -> User:
    """Extract Supabase user id from JWT.
    
    NOTE: For development, this skips signature verification.
    In production, validate using Supabase's JWKS or JWT secret.
    """
    if credentials is None or credentials.scheme.lower() != "bearer":
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Missing or invalid Authorization header",
        )

    token = credentials.credentials

    try:
        # Dev mode: decode without verification
        claims = jwt.get_unverified_claims(token)
        user_id = claims.get("sub")
        
        if not user_id:
            raise HTTPException(
                status_code=status.HTTP_401_UNAUTHORIZED,
                detail="Token missing subject",
            )

        return User(id=user_id)

    except HTTPException:
        raise
    except Exception:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid token",
        )


async def get_optional_user(
    credentials: HTTPAuthorizationCredentials = Depends(bearer_scheme),
) -> User | None:
    """Get user if authenticated, None otherwise (for dev bypass)."""
    if settings.DEV_BYPASS_AUTH:
        return None
    
    if credentials is None:
        return None
    
    try:
        return await get_current_user(credentials)
    except HTTPException:
        return None
