_archived/backend_archived/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-jose[cryptography]
greenlet
pydantic-settings
httpx
python-dotenv
supabase
